+++ pyobjc-framework-Security/Modules/_Security.m
@@ -466,9 +466,66 @@ parse_itemset(PyObject* value, AuthorizationItemSet* itemset)
         }
 
         for (i = 0; i < PyTuple_GET_SIZE(seq); i++) {
-            if (PyObjC_PythonToObjC("{_AuthorizationItem=^cL^vI}",
-                                    PyTuple_GET_ITEM(seq, i), itemset->items + i)
-                < 0) {
+            PyObject* cur_seq = PySequence_Tuple(PyTuple_GET_ITEM(seq, i));
+            if (cur_seq == NULL) {
+                PyErr_Format(PyExc_TypeError, "items[%ld] is not a sequence", (long)i);
+                PyMem_Free(itemset->items);
+                return 0;
+            }
+            if (PyTuple_GET_SIZE(cur_seq) != 4) {
+                PyErr_Format(PyExc_TypeError, "items[%ld] is not a sequence of 4 items",
+                             (long)i);
+                PyMem_Free(itemset->items);
+                return 0;
+            }
+
+            if (!PyBytes_Check(PyTuple_GET_ITEM(cur_seq, 0))) {
+                PyErr_Format(PyExc_TypeError, "items[%ld].name is not a byte string",
+                             (long)i);
+                PyMem_Free(itemset->items);
+                return 0;
+            }
+
+            itemset->items[i].name = PyBytes_AsString(PyTuple_GET_ITEM(cur_seq, 0));
+
+            if (PyObjC_PythonToObjC(@encode(size_t), PyTuple_GET_ITEM(cur_seq, 1),
+                                    &(itemset->items[i].valueLength))
+                == -1) {
+                PyErr_Format(PyExc_TypeError, "items[%ld].valueLength is not an integer",
+                             (long)i);
+                PyMem_Free(itemset->items);
+                return 0;
+            }
+
+            if (PyTuple_GET_ITEM(cur_seq, 2) == Py_None) {
+                if (itemset->items[i].valueLength != 0) {
+                    PyErr_Format(PyExc_TypeError,
+                                 "items[%ld].value is None, valueLength != 0", (long)i);
+                    PyMem_Free(itemset->items);
+                    return 0;
+                }
+
+                itemset->items[i].value = NULL;
+
+            } else {
+                if (!PyBytes_Check(PyTuple_GET_ITEM(cur_seq, 2))
+                    || PyBytes_Size(PyTuple_GET_ITEM(cur_seq, 2))
+                           != (Py_ssize_t)itemset->items[i].valueLength) {
+                    PyErr_Format(PyExc_TypeError,
+                                 "items[%ld].value is not a byte string of length %ld",
+                                 (long)i, (long)(itemset->items[i].valueLength));
+                    PyMem_Free(itemset->items);
+                    return 0;
+                }
+
+                itemset->items[i].value = PyBytes_AsString(PyTuple_GET_ITEM(cur_seq, 2));
+            }
+
+            if (PyObjC_PythonToObjC(@encode(UInt32), PyTuple_GET_ITEM(cur_seq, 3),
+                                    &(itemset->items[i].flags))
+                == -1) {
+                PyErr_Format(PyExc_TypeError, "items[%ld].flags is not an integer",
+                             (long)i);
                 PyMem_Free(itemset->items);
                 return 0;
             }
@@ -477,32 +534,88 @@ parse_itemset(PyObject* value, AuthorizationItemSet* itemset)
     return 1;
 }
 
-static PyObject*
-build_itemset(AuthorizationItemSet* itemset)
+static void
+free_itemset(AuthorizationItemSet* itemset)
 {
-    PyObject* result;
+    PyMem_Free(itemset->items);
+}
 
-    if (itemset == NULL) {
-        Py_INCREF(Py_None);
-        return Py_None;
+static PyObject*
+build_itemset(AuthorizationItemSet* _Nonnull itemset)
+{
+    PyObject* result = NULL;
+    PyObject* t      = NULL;
+    PyObject* o      = NULL;
 
-    } else {
-        UInt32 i;
-        result = PyTuple_New(itemset->count);
-        if (result == NULL) // LCOV_BR_EXCL_LINE
-            return NULL;    // LCOV_EXCL_LINE
+    assert(itemset != NULL);
 
-        for (i = 0; i < itemset->count; i++) {
-            PyObject* t =
-                PyObjC_ObjCToPython("{_AuthorizationItem=^cL^vI}", itemset->items + i);
-            if (t == NULL) {
-                Py_DECREF(result);
-                return NULL;
-            }
-            PyTuple_SET_ITEM(result, i, t);
+    UInt32 i;
+    result = PyTuple_New(itemset->count);
+    if (result == NULL) // LCOV_BR_EXCL_LINE
+        return NULL;    // LCOV_EXCL_LINE
+
+    for (i = 0; i < itemset->count; i++) {
+        Py_ssize_t         packed = -1;
+        AuthorizationItem* item   = itemset->items + i;
+
+        t = PyObjC_CreateRegisteredStruct("{_AuthorizationItem=^cL^vI}",
+                                          sizeof("{_AuthorizationItem=^cL^vI}") - 1, NULL,
+                                          &packed);
+        if (t == NULL)  // LCOV_BR_EXCL_LINE
+            goto error; // LCOV_EXCL_LINE
+
+        o = PyBytes_FromString(item->name);
+        if (o == NULL)  // LCOV_BR_EXCL_LINE
+            goto error; // LCOV_EXCL_LINE
+
+        if (PySequence_SetItem(t, 0, o) == -1) // LCOV_BR_EXCL_LINE
+            goto error;                        // LCOV_EXCL_LINE
+
+        Py_CLEAR(o);
+
+        o = PyLong_FromUnsignedLong(item->valueLength);
+        if (o == NULL)  // LCOV_BR_EXCL_LINE
+            goto error; // LCOV_EXCL_LINE
+
+        if (PySequence_SetItem(t, 1, o) == -1) // LCOV_BR_EXCL_LINE
+            goto error;                        // LCOV_EXCL_LINE
+
+        Py_CLEAR(o);
+
+        if (item->value == NULL) {
+            o = Py_None;
+            Py_INCREF(Py_None);
+        } else {
+            o = PyBytes_FromStringAndSize(item->value, item->valueLength);
+            if (o == NULL)  // LCOV_BR_EXCL_LINE
+                goto error; // LCOV_EXCL_LINE
         }
+
+        if (PySequence_SetItem(t, 2, o) == -1) // LCOV_BR_EXCL_LINE
+            goto error;                        // LCOV_EXCL_LINE
+
+        Py_CLEAR(o);
+
+        o = PyLong_FromUnsignedLong(item->flags);
+        if (o == NULL)  // LCOV_BR_EXCL_LINE
+            goto error; // LCOV_EXCL_LINE
+
+        if (PySequence_SetItem(t, 3, o) == -1) // LCOV_BR_EXCL_LINE
+            goto error;                        // LCOV_EXCL_LINE
+
+        Py_CLEAR(o);
+
+        PyTuple_SET_ITEM(result, i, t);
     }
     return result;
+
+error:
+    // LCOV_EXCL_START
+    Py_CLEAR(result);
+    Py_CLEAR(t);
+    Py_CLEAR(o);
+    return NULL;
+    // LCOV_EXCL_STOP
 }
 
 static PyObject*
@@ -554,8 +667,8 @@ m_AuthorizationCreate(PyObject* meth, PyObject* _Nonnull const* _Nonnull args,
         }
     Py_END_ALLOW_THREADS
 
-    PyMem_Free(rights.items);
-    PyMem_Free(environment.items);
+    free_itemset(&rights);
+    free_itemset(&environment);
 
     if (PyErr_Occurred()) // LCOV_BR_EXCL_LINE
         return NULL;      // LCOV_EXCL_LINE
@@ -668,8 +781,8 @@ m_AuthorizationCopyRights(PyObject* meth, PyObject* _Nonnull const* _Nonnull arg
         }
     Py_END_ALLOW_THREADS
 
-    PyMem_Free(rights.items);
-    PyMem_Free(environment.items);
+    free_itemset(&rights);
+    free_itemset(&environment);
 
     if (PyErr_Occurred()) // LCOV_BR_EXCL_LINE
         return NULL;      // LCOV_EXCL_LINE
@@ -753,12 +866,14 @@ m_AuthorizationCopyRightsAsync(PyObject* meth, PyObject* _Nonnull const* _Nonnul
                       py_result = PyObject_CallFunction(py_callback, "iO", err,
                                                         py_authorizedRights);
                       if (py_result == NULL) {
-                          PyObjC_GIL_FORWARD_EXC();
+                          /* Don't raise as ObjC exception, will cause hard crash */
+                          PyErr_WriteUnraisable(py_callback);
                       } else if (py_result != Py_None) {
                           Py_DECREF(py_result);
                           PyErr_SetString(PyExc_TypeError,
                                           "callbackBlock returned value");
-                          PyObjC_GIL_FORWARD_EXC();
+                          /* Don't raise as ObjC exception, will cause hard crash */
+                          PyErr_WriteUnraisable(py_callback);
                       } else {
                           Py_DECREF(py_result);
                       }